Ensure we are Monitoring Instance Memory (#2)
* Ensure we are Monitoring Instance Memory

Cloudwatch doesn't monitor instance memory by default.
We need to pull in a custom script and add it to a cron (which will come
in a future PR). Ensure that this is present in the /scripts directory
for now.

* Download CloudWatch memory monitoring scripts over https

* Improve status check for presence of CloudWatch memory logging directory

Don't just check that the scripts directory exists, but the actual
CloudWatch memory script directory within it before creating or moving
on.